[Tosa] Add support for legalizing scatter_nd
This commit legalizes tf(l).scatter_nd to tosa.scatter in a similar
approach to the existing gather_nd support. Specifically, inputs are
rewritten to the expected formats by tosa.scatter and finally the
output of tosa.scatter is reshaped to the expected output of scatter_nd.

tosa.scatter does not support duplicated indices while TF does.
Therefore, we restrict legalization of scatter_nd to a constant indices
tensor only and check that the provided indices are unique.

Signed-off-by: Luke Hutton <luke.hutton@arm.com>
Change-Id: Id0c6c4df1aa09807bc8f47c430924846f4670ba8